backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
SQLAlchemy<2.0
Flask-JWT-Extended==4.4.4
python-dotenv==1.0.0
werkzeug==2.3.7
